Change all parseInts to parseFloat
The current parseInt doesn't handle decimals, thus rounds all of
the numbers to next integer, which we don't want. To fix this, all
instances of parseInt are changed to parseFloat

+++ script.js
@@ -71,17 +71,17 @@ $(function() {
     })
     $("#equals").click(function(){
         if (operator === "+"){
-            number = (parseInt(number, 10) + parseInt(newnumber,10)).toString(10);
+            number = (parseFloat(number, 10) + parseFloat(newnumber,10)).toString(10);
         } else if (operator === "-"){
-            number = (parseInt(newnumber, 10) - parseInt(number,10)).toString(10);
+            number = (parseFloat(newnumber, 10) - parseFloat(number,10)).toString(10);
         } else if (operator === "/"){
-            number = (parseInt(newnumber, 10) / parseInt(number,10)).toString(10);
+            number = (parseFloat(newnumber, 10) / parseFloat(number,10)).toString(10);
         } else if (operator === "*"){
-            number = (parseInt(newnumber, 10) * parseInt(number,10)).toString(10);
+            number = (parseFloat(newnumber, 10) * parseFloat(number,10)).toString(10);
         } else if (operator === "&sqrt;"){
-            number = (Math.sqrt(parseInt(number, 10))).toString(10);   
+            number = (Math.sqrt(parseFloat(number, 10))).toString(10);   
         } else if (operator === "^"){
-            number = (Math.pow(parseInt(newnumber, 10))).toString(10);   
+            number = (Math.pow(parseFloat(newnumber, 10))).toString(10);   
         }
         totaldiv.text(number);
         testNumLength(number);
